Move generator class to Tram::Policy
