+++ lib/generators/tram/policy/templates/policy_class.erb
