+++ lib/generators/tram/policy/templates/policy_class_spec.erb
